peer, netlink: cast via intptr_t
Cast void* to int (and vice-versa) via intptr_t; not doing this
produces warning (promoted to error via -Werror) on 64-bit systems.

+++ nhrp/nhrp_peer.c
@@ -1355,7 +1355,7 @@ int nhrp_peer_remove_matching(void *ctx, struct nhrp_peer *peer)
 
 int nhrp_peer_set_used_matching(void *ctx, struct nhrp_peer *peer)
 {
-	int used = (int) ctx;
+	int used = (int) (intptr_t) ctx;
 
 	if (used) {
 		peer->flags |= NHRP_PEER_FLAG_USED;
